server/api/requirements.txt
asgiref==3.7.2
coverage==7.4.4
dj-database-url==2.1.0
Django==5.0.3
django-cors-headers==4.3.1
django-environ==0.11.2
django-rest-framework==0.1.0
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-decouple==3.8
sqlparse==0.4.4
typing_extensions==4.10.0
